moved init_rovers_and_cameras to models module; rearranged tests accordingly.
