+++ mars_street_view/models.py
@@ -203,3 +203,60 @@ class Camera(Base):
     rover_name = Column(String, ForeignKey('rovers.name'))
     photos = relationship('Photo', back_populates='camera', lazy='dynamic')
     rover = relationship('Rover', back_populates='cameras')
+
+
+CAMERAS = {
+    'FHAZ': "Front Hazard Avoidance Camera",
+    "NAVCAM": "Navigation Camera",
+    "MAST": "Mast Camera",
+    "CHEMCAM": "Chemistry and Camera Complex",
+    "MAHLI": "Mars Hand Lens Imager",
+    "MARDI": "Mars Descent Imager",
+    "RHAZ": "Rear Hazard Avoidance Camera",
+    "PANCAM": "Panoramic Camera",
+    "MINITES": "Miniature Thermal Emission Spectrometer (Mini-TES)",
+    "ENTRY": "Entry, Descent, and Landing Camera"
+}
+
+
+ROVERS = [
+    {'name': 'Curiosity',
+     'landing_date': "2012-08-06",
+     'cameras': ['FHAZ', "NAVCAM", "MAST", "CHEMCAM", "MAHLI", "MARDI",
+                 "RHAZ"],
+     'max_date': "2016-03-28",
+     'max_sol': 1295,
+     'total_photos': 246346,
+     },
+    {'name': 'Spirit',
+     'landing_date': "2004-01-04",
+     'cameras': ['FHAZ', "NAVCAM", "RHAZ", "PANCAM", "MINITES", "ENTRY"],
+     'max_date': "2010-03-21",
+     'max_sol': 2208,
+     'total_photos': 124550,
+     },
+    {'name': 'Opportunity',
+     'landing_date': "2004-01-25",
+     'max_date': "2016-03-28",
+     'cameras': ['FHAZ', "NAVCAM", "RHAZ", "PANCAM", "MINITES", "ENTRY"],
+     'max_sol': 4328,
+     'total_photos': 178933,
+     }
+]
+
+
+def init_rovers_and_cameras():
+    """Create all Rovers and Cameras and save in database."""
+    objects_list = []
+    for rover_dict in ROVERS:
+        objects_list.append(Rover(**rover_dict))
+        rover_name = rover_dict['name']
+        cam_list = rover_dict['cameras']
+        for short_name in cam_list:
+            cam_dict = {
+                'name': short_name,
+                'rover_name': rover_name,
+                'full_name': CAMERAS[short_name]
+            }
+            objects_list.append(Camera(**cam_dict))
+    return objects_list
